Fixed bug with outer radius calculation.

+++ include/EvolutionAnalyser.h
@@ -32,12 +32,12 @@ public:
 
 private:
   struct Record {
-    float time = 0.0f;
+    double time = 0.0;
     float max_dens = 0.0f;
     float max_temp = 0.0f;
     float disc_mass = 0.0f;
     float star_mass = 0.0f;
-    float r_out[3] = {0.0f, 0.0f, 0.0f};
+    double r_out[3] = {0.0, 0.0, 0.0};
   };
 
   NameData mNameData;
